chore: Update package.json dependencies
This commit updates the dependencies in the package.json file. It adds the "axios" package with version "^1.6.5" and updates the versions of "dotenv", "express", "mongodb", "mongoose", and "nodemon" packages.

Note: This is a suggested commit message based on the code changes provided. Please modify it as needed to adhere to your repository's commit message conventions.